MS-939 Sync info observer calculation use case extracted from the viewmodel
